More work on ANTLR3.4 revision and XCode4 LLVM2.0 compiler
[git-p4: depot-paths = "//depot/code/antlr/antlr3-main/": change = 8468]

+++ runtime/ObjC/Framework/ANTLRLexerState.h
@@ -36,19 +36,19 @@
 - (id<ANTLRToken>) getToken;
 - (void) setToken:(id<ANTLRToken>) theToken;
 
-- (unsigned int) getTokenType;
+- (NSUInteger) getTokenType;
 - (void) setTokenType:(unsigned int) theTokenType;
 
-- (unsigned int) getChannel;
+- (NSUInteger) getChannel;
 - (void) setChannel:(unsigned int) theChannel;
 
-- (unsigned int) getTokenStartLine;
+- (NSUInteger) getTokenStartLine;
 - (void) setTokenStartLine:(unsigned int) theTokenStartLine;
 
-- (unsigned int) getTokenCharPositionInLine;
+- (NSUInteger) getTokenCharPositionInLine;
 - (void) setTokenCharPositionInLine:(unsigned int) theCharPosition;
 
-- (int) getTokenStartCharIndex;
+- (NSInteger) getTokenStartCharIndex;
 - (void) setTokenStartCharIndex:(int) theTokenStartCharIndex;
 
 - (NSString *) text;
